Refactored messaging
