+++ module.json
@@ -2,7 +2,7 @@
   "name": "vtta-dndbeyond",
   "title": "Virtual Tabletop Assets - D&D Beyond Integration",
   "description": "Bringing your licensed content from D&D Beyond into Foundry VTT",
-  "version": "2.2.2",
+  "version": "2.2.3",
   "minimumCoreVersion": "0.5.5",
   "compatibleCoreVersion": "0.5.5",
   "author": "Sebastian Will",
