+++ src/messaging/type/_addEntity.js
