+++ src/utils.js
@@ -314,6 +314,7 @@ let utils = {
     }
     return undefined;
   },
+
   uploadImage: async function (url, targetDirectory, baseFilename) {
     async function download(url) {
       return new Promise((resolve, reject) => {
@@ -414,6 +415,7 @@ let utils = {
       return null;
     }
   },
+
   getFolder: async (kind, type = "", race = "") => {
     let getOrCreateFolder = async (root, entityType, folderName) => {
       const baseColor = "#98020a";
@@ -489,6 +491,85 @@ let utils = {
       return entityFolder;
     }
   },
+
+  normalizeString: (str) => {
+    return str.toLowerCase().replace(/\W/g, "");
+  },
+
+  /**
+   * Queries a compendium for a given entity name
+   * @returns the index entries of all matches, otherwise an empty array
+   */
+  queryCompendium: async (compendiumName, entityName) => {
+    entityName = utils.normalizeString(entityName);
+
+    let compendium = game.packs.find(
+      (pack) => pack.collection === compendiumName
+    );
+    if (!compendium) return null;
+    let index = await compendium.getIndex();
+    let entity = index.find(
+      (entity) => utils.normalizeString(entity.name) === entityName
+    );
+    return entity;
+  },
+
+  /**
+   * Creates or updates a given entity
+   */
+  createCompendiumEntry: async (
+    compendiumName,
+    entity,
+    updateExistingEntry = false
+  ) => {
+    let compendium = game.packs.find(
+      (pack) => pack.collection === compendiumName
+    );
+
+    if (!compendium) return null;
+
+    let existingEntry = await utils.queryCompendium(
+      compendiumName,
+      entity.name
+    );
+    if (existingEntry) {
+      console.log("Entry exists already:");
+      console.log(existingEntry);
+      if (updateExistingEntry) {
+        // update all existing entries
+        existingEntry = await compendium.updateEntity({
+          ...entity.data,
+          _id: existingEntry._id,
+        });
+
+        return {
+          _id: existingEntry._id,
+          img: existingEntry.img,
+          name: existingEntry.name,
+        };
+      } else {
+        console.log("Update: no");
+        return existingEntry;
+      }
+    } else {
+      console.log("Entry does not exist");
+      let compendiumEntry = await compendium.createEntity(entity.data);
+      console.log(compendiumEntry);
+      return {
+        _id: compendiumEntry._id,
+        img: compendiumEntry.img,
+        name: compendiumEntry.name,
+      };
+    }
+  },
+
+  getFolderHierarchy: (folder) => {
+    if (!folder || !folder._parent) return "/";
+    return folder._parent._id !== null
+      ? `${utils.getFolderHierarchy(folder._parent)}/${folder.name}`
+      : `/${folder.name}`;
+  },
+
   log: (msg, section = "general") => {
     const LOG_PREFIX = "VTTA D&D Beyond";
     if (
